Refactor Monthly Reviews to Timesheets
Replaced the "Monthly Reviews" concept with "Timesheets" across the codebase. This includes renaming entities, services, controllers, and tests to reflect the new terminology.

- Removed `MonthlyReview` entity and introduced `Timesheet` entity.
- Added `TimesheetsController` to handle API endpoints.
- Replaced `MonthlyReviewService` with `TimesheetService` for business logic.
- Updated `OneTimeContext` to include `Timesheets` and remove `MonthlyReviews`.
- Introduced `SubmitTimesheetDto` and `TimesheetDto` for data transfer.
- Updated `TimeEntry` and `Project` entities to store `Status` as an integer.
- Added hierarchical user relationships via `ManagerId` in `JNUser`.
- Updated dependency injection in `Program.cs` to register `TimesheetService`.
- Updated `efpt.config.json` to reflect database schema changes.
- Added `TimesheetServiceTests` and `TimesheetsTests` to ensure functionality.
- Updated `TimeEntryConverter` to handle `Status` casting between int and enum.

This refactor ensures the application now operates with "Timesheets" as the primary entity for managing user submissions within specified periods.

+++ src/backend/OneTime.Core.Tests/TestData/MockData.cs
@@ -27,7 +27,7 @@ namespace OneTime.Core.Tests.TestData
                 ProjectId = projectId,
                 Hours = hours,
                 Date = default,
-                Status = TimeEntryStatus.Pending
+                Status = (int)TimeEntryStatus.Pending
             };
         }
     }
